+++ tests/examples/make_inventory_list/plan.py
@@ -4,4 +4,4 @@ from stepup.core.api import static
 from stepup.reprep.api import make_inventory
 
 static("README.md")
-make_inventory(["README.md"], "inventory.txt")
+make_inventory("README.md", "inventory.txt")