+++ tests/examples/zip_tree/plan.py
@@ -10,5 +10,5 @@ glob("data/**/", _defer=True)
 out_files = glob("data/**/*.out")
 
 # Create an inventory file with all .out files and zip it.
-make_inventory(out_files, "inventory.txt")
+make_inventory(*out_files, "inventory.txt")
 zip_inventory("inventory.txt", "upload.zip")