+++ tests/test_flatten_latex.py
@@ -71,3 +71,23 @@ def test_latex_flat(path_tmp):
     with open(flatdir / "main.tex") as fh:
         result = fh.read()
     assert result.strip() == EXPECTED.strip()
+
+
+def test_latex_flat_simple(path_tmp):
+    with open(path_tmp / "main.tex", "w") as fh:
+        fh.write("Not so much to see here.")
+    with open(path_tmp / "flat.tex", "w") as fh:
+        flatten_latex(path_tmp / "main.tex", fh, path_tmp)
+    with open(path_tmp / "flat.tex") as fh:
+        result = fh.read()
+    assert result.strip() == "Not so much to see here."
+
+
+def test_latex_flat_empty(path_tmp):
+    with open(path_tmp / "main.tex", "w") as fh:
+        pass
+    with open(path_tmp / "flat.tex", "w") as fh:
+        flatten_latex(path_tmp / "main.tex", fh, path_tmp)
+    with open(path_tmp / "flat.tex") as fh:
+        result = fh.read()
+    assert result.strip() == ""